- rename compat to compat.hadoop - Fix FindBugs error around non-static inner classes
