+++ gobblin-api/src/main/java/gobblin/compat/hadoop/TextSerializer.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package gobblin.compat;
+package gobblin.compat.hadoop;
 
 import java.io.DataInput;
 import java.io.DataOutput;
